marta blog final

+++ MartaBlog/src/main/java/com/sanluis/spring/MartaBlog/repository/PostDAO.java
@@ -6,9 +6,9 @@ import com.sanluis.spring.MartaBlog.vo.Post;
 
 public interface PostDAO {
 	public List<Post> getAll();
-	public List<Post> getAllByidAutor(Long id);
-	public Post findById(Long id);
+	public List<Post> getAllByidAutor(int id);
+	public Post findById(int id);
 	public void insert(Post post);
 	public void update(Post post);
-	public void deleteById(Long id);
+	public void deleteById(int id);
 }
