+++ MartaBlog/src/main/java/com/sanluis/spring/MartaBlog/service/PostService.java
@@ -7,9 +7,9 @@ import com.sanluis.spring.MartaBlog.vo.Post;
 
 public interface PostService {
 	public List<Post> getAll();
-	public List<Post> getAllByidAutor(Long id);
-	public Post findById(Long id);
+	public List<Post> getAllByidAutor(int id);
+	public Post findById(int id);
 	public void insert(Post post);
 	public void update(Post post);
-	public void deleteById(Long id);
+	public void deleteById(int id);
 }
